# packages/protocol/src/generated/openapi.bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComaMessenger API",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/chats/{chat_id}/messages": {
      "get": {
        "operationId": "listMessages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChatId"
          },
          {
            "name": "before_seq",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "thread_root_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages visible to the current chat member, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createMessage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChatId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The original message for an idempotent replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "201": {
            "description": "Message created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}/pins": {
      "get": {
        "operationId": "listMessagePins",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChatId"
          }
        ],
        "responses": {
          "200": {
            "description": "Current pins in the chat, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePinList"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/messages/{message_id}": {
      "patch": {
        "operationId": "updateMessage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteMessage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "200": {
            "description": "Soft-deleted message, or the existing tombstone for an idempotent replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/messages/{message_id}/reactions/{emoji}": {
      "put": {
        "operationId": "putReaction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          },
          {
            "name": "emoji",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Existing reaction for an idempotent replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reaction"
                }
              }
            }
          },
          "201": {
            "description": "Reaction added.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reaction"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteReaction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          },
          {
            "name": "emoji",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Reaction is absent."
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/reactions": {
      "get": {
        "operationId": "listReactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "200": {
            "description": "Current reactions on the message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionList"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/pin": {
      "put": {
        "operationId": "putMessagePin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "200": {
            "description": "Existing pin for an idempotent replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePin"
                }
              }
            }
          },
          "201": {
            "description": "Message pinned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePin"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteMessagePin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "204": {
            "description": "Message pin is absent."
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/forward": {
      "post": {
        "operationId": "forwardMessage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Original forwarded snapshot for an idempotent replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "201": {
            "description": "Forwarded snapshot created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/thread/follow": {
      "put": {
        "operationId": "followThread",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "200": {
            "description": "Existing thread follow for an idempotent replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadFollow"
                }
              }
            }
          },
          "201": {
            "description": "Thread followed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadFollow"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "unfollowThread",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/MessageId"
          }
        ],
        "responses": {
          "204": {
            "description": "Thread follow is absent."
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/threads": {
      "get": {
        "operationId": "listFollowedThreads",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "before_seq",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Followed threads visible to the current member, ordered by latest activity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadPage"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Error": {
        "description": "Stable API error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "parameters": {
      "ChatId": {
        "name": "chat_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "MessageId": {
        "name": "message_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "already_bootstrapped",
          "chat_conflict",
          "chat_not_found",
          "forbidden",
          "idempotency_conflict",
          "internal_error",
          "invalid_credentials",
          "invalid_refresh_token",
          "invalid_request",
          "invitation_invalid",
          "message_not_found",
          "origin_not_allowed",
          "payload_too_large",
          "rate_limited",
          "service_not_ready",
          "session_not_found",
          "unauthorized",
          "unsupported_format",
          "validation_failed",
          "version_conflict"
        ]
      },
...
      "Message": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "chat_id",
          "actor_id",
          "client_msg_id",
          "type",
          "body",
          "body_format",
          "mentioned_actor_ids",
          "version",
          "created_seq",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "chat_id": {
            "type": "string",
            "format": "uuid"
          },
          "actor_id": {
            "type": "string",
            "format": "uuid"
          },
          "client_msg_id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "text"
            ]
          },
          "body": {
            "type": "string"
          },
          "body_format": {
            "type": "string",
            "enum": [
              "plain",
              "markdown"
            ]
          },
          "reply_to_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "thread_root_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "created_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "edited_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "forwarded_from": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ForwardAttribution"
              },
              {
                "type": "null"
              }
            ]
          },
          "mentioned_actor_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          }
        }
      },
      "MessagePage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "messages",
          "next_before_seq"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "next_before_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 1
          }
        }
      },
...
      "CreateMessageRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "client_msg_id",
          "body"
        ],
        "properties": {
          "client_msg_id": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string",
            "minLength": 1
          },
          "body_format": {
            "type": "string",
            "enum": [
              "plain",
              "markdown"
            ],
            "default": "plain"
          },
          "reply_to_id": {
            "type": "string",
            "format": "uuid"
          },
          "thread_root_id": {
            "type": "string",
            "format": "uuid"
          },
          "mentioned_actor_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          }
        }
      },
      "UpdateMessageRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "body",
          "expected_version"
        ],
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1
          },
          "body_format": {
            "type": "string",
            "enum": [
              "plain",
              "markdown"
            ],
            "default": "plain"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "mentioned_actor_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          }
        }
      },
      "ForwardAttribution": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "author_name",
          "author_handle",
          "created_at"
        ],
        "properties": {
          "author_name": {
            "type": "string"
          },
          "author_handle": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ForwardMessageRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "chat_id",
          "client_msg_id"
        ],
        "properties": {
          "chat_id": {
            "type": "string",
            "format": "uuid"
          },
          "client_msg_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Reaction": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message_id",
          "actor_id",
          "emoji",
          "created_at"
        ],
        "properties": {
          "message_id": {
            "type": "string",
            "format": "uuid"
          },
          "actor_id": {
            "type": "string",
            "format": "uuid"
          },
          "emoji": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReactionList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reactions"
        ],
        "properties": {
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reaction"
            }
          }
        }
      },
      "MessagePin": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message_id",
          "pinned_by",
          "pinned_at"
        ],
        "properties": {
          "message_id": {
            "type": "string",
            "format": "uuid"
          },
          "pinned_by": {
            "type": "string",
            "format": "uuid"
          },
          "pinned_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MessagePinList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "pins"
        ],
        "properties": {
          "pins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessagePin"
            }
          }
        }
      },
      "ThreadFollow": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "thread_root_id",
          "followed_at"
        ],
        "properties": {
          "thread_root_id": {
            "type": "string",
            "format": "uuid"
          },
          "followed_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ThreadPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "threads",
          "next_before_seq"
        ],
        "properties": {
          "threads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadSummary"
            }
          },
          "next_before_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 1
          }
        }
      },
...
    }
  }
}